Fixed version of MLCC carrier fighter script.
If we didn't configure a task force size for a ship class we'd
end up sitting there until the enemy ship left or someone else
destroyed it. Instead of doing that, which is kind of stupid, just
throw everything at it.

The classes are TL, TM, TP, TS and drones. With drones being the most
annoying (the others usually leave).