Repair carrier script.
I've no idea what's going on. First, we used a $Distance that wasn't set
until later, although it should have roughly the right value after a
while, but fixing that didn't help. Apparently if we do 'follow object'
with a distance of 40k it just immediately returns FLT_TARGETREACHED.
Just set the distance to 15k if it's above 15k (maybe more distance
works, but 15k should be good enough since we won't get too close anyway
before the timeout hits and we try this again).